Use current URL if empty string is used for graphql_path
I would like to serve GraphiQL from the same path as my GraphQL
endpoint. (GraphiQL is returned for GETs, and query results are returned
for POSTs.)

I need this feature, because I'd like to package both GraphQL and
GraphiQL into a mountable engine (in a gem), and I won't actually know
the full graphql_path at the time that my engine's routes are drawn.
(It's determined by where my gem's engine is mounted by the Rails app.)

This PR allows use of an empty string to indicate the the GraphQL
endpoint is the same URL that GraphiQL loaded from. I've added
documentation for this minor change.