Exposing server error in response pane if invalid JSON returned
If a server 500s or otherwise fails to return valid JSON, the GraphiQL
interface will simply show the error of trying to parse the response as
JSON. This means that for any meaningful GraphQL development, you have
to keep the network tab open at all times to read the actual server
response.

This replaces `.json()` with `.text()` and `JSON.parse()`. This is a
quick fix without much thought put into it, but I couldn't find a more
obvious way to preserve the JSON parsing error along with the actual
server error. The Response() API is...interesting.